got the first brush strokes representation working. The representation uses an array of points, creates quads around those points, and uses the genes of those points to create brush strokes. The initial creation of the genes is done in the CPU and the putting everything together is done in vertex shader.
Next step is: fix the culling order, so that it is either front or
back facing.
Move draw now to command buffers to gain more control over
exectuation time.
Create multiplie compute buffers and rendering for different
memeber of population.
create fitness function through compute shaders that
compare the result to the picture.